fix: alerts for deletion results
