+++ packages/main/package.json
@@ -21,7 +21,6 @@
     "@mui/styles": "^5.13.2",
     "@szhsin/react-menu": "^4.0.0",
     "@tanstack/match-sorter-utils": "^8.8.4",
-    "@tanstack/react-query": "^4.29.25",
     "@tanstack/react-table": "^8.9.2",
     "axios": "^1.4.0",
     "classnames": "^2.3.2",
@@ -78,6 +77,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
     "happy-dom": "^9.20.3",
+    "prettier": "^3.0.3",
     "sass": "^1.63.6",
     "scss": "^0.2.4",
     "typescript": "^5.0.2",
