+++ packages/main/plugins/Cardinality/SeriesGroup.tsx
@@ -1,66 +1,33 @@
-import React, { useState, useMemo } from "react";
-import { SeriesHeader, SeriesHeaderProps } from "./SeriesHeader";
-import { SeriesRow, SeriesRowHeaders, SeriesRowProps } from "./SeriesRow";
+import React from "react";
+import SeriesHeader, { type SeriesHeaderProps } from "./SeriesHeader";
+import SeriesRow, { type SeriesRowProps } from "./SeriesRow";
+import SeriesRowHeaders from "./SeriesRowHeaders";
 import useTheme from "@ui/theme/useTheme";
-import { cx, css } from "@emotion/css";
-import { sortByCol} from './helpers'
-
-export const SeriesGroupContainer = (theme: any) => css`
-    margin: 4px;
-    display: flex;
-    flex-direction: column;
-    padding: 8px 4px;
-    background: ${theme.shadow};
-    .c-header {
-        font-size: 14px;
-        padding: 8px 6px;
-        border-bottom: 1px solid ${theme.neutral};
-       
-        font-weight: bold;
-    }
-    .c-table {
-        display: table;
-        widht: 100%;
-    }
-`;
+import { useSeriesGroupStyles } from "./SeriesGroupStyles";
+import { useSortedColumns } from "./useSortColumn";
 
 export type SeriesGroupProps = {
     rows: SeriesRowProps[];
     sectionHeader: string;
-    sectionHeaderName:string;
+    sectionHeaderName: string;
 } & SeriesHeaderProps;
 
-// format the header according to the header type
-
+// This components is used to display a group of series
 export const SeriesGroup: React.FC<SeriesGroupProps> = ({
     title,
     rows,
     sectionHeader,
-    sectionHeaderName
+    sectionHeaderName,
 }: SeriesGroupProps) => {
     const theme = useTheme();
 
-    const [sortColumn, setSortColumn] = useState({
-        col: "value",
-        order: "desc",
-    });
+    const { seriesGroupContainer, seriesGroupStyles } =
+        useSeriesGroupStyles(theme);
 
-    const sortedRows = useMemo(() => {
-        const { col, order } = sortColumn;
-        return sortByCol(rows, col, order);
-    }, [rows, sortColumn]);
-
-    
-
-    const handleSort = (e: any) => {
-        setSortColumn((prev: any) => ({
-            col: e,
-            order: prev.order === "asc" ? "desc" : "asc",
-        }));
-    };
+    const { sortedRows, handleSort } = useSortedColumns(rows);
 
     return (
-        <div className={cx(SeriesGroupContainer(theme))}>
+        <div className={seriesGroupContainer}>
             <SeriesHeader title={title} />
             <div className="c-table">
                 {rows && (
@@ -68,13 +35,18 @@ export const SeriesGroup: React.FC<SeriesGroupProps> = ({
                         handleSort={handleSort}
                         headerName={sectionHeaderName}
                         name={sectionHeader}
-                        theme={theme}
+                        theme={seriesGroupStyles}
                     />
                 )}
                 {sortedRows &&
                     sortedRows?.length > 0 &&
                     sortedRows.map((row: SeriesRowProps, key: number) => (
-                        <SeriesRow  key={key} headerName={sectionHeaderName} {...row} />
+                        <SeriesRow
+                            key={key}
+                            theme={seriesGroupStyles}
+                            hasShare={sectionHeaderName !== "labelValueCountByLabelName"}
+                            {...row}
+                        />
                     ))}
             </div>
         </div>
