+++ packages/main/plugins/Cardinality/SeriesHeader.tsx
@@ -3,7 +3,10 @@ import React from 'react'
 export type SeriesHeaderProps = {
     title: string;
 };
+// here should be the tabs for table / chart view
 
-export const SeriesHeader:React.FC<SeriesHeaderProps> = ({ title }: SeriesHeaderProps) => {
+const SeriesHeader:React.FC<SeriesHeaderProps> = ({ title }: SeriesHeaderProps) => {
     return <div className="c-header">{title}</div>;
 };
+
+export default SeriesHeader;
